+++ exporter/datadogexporter/internal/model/attributes/gcp/gcp.go
@@ -17,7 +17,7 @@ import (
 	"fmt"
 
 	"go.opentelemetry.io/collector/model/pdata"
-	conventions "go.opentelemetry.io/collector/model/semconv/v1.5.0"
+	conventions "go.opentelemetry.io/collector/model/semconv/v1.6.1"
 )
 
 type HostInfo struct {