+++ exporter/elasticexporter/internal/translator/elastic/metadata.go
@@ -21,7 +21,7 @@ import (
 	"go.elastic.co/apm/model"
 	"go.elastic.co/fastjson"
 	"go.opentelemetry.io/collector/model/pdata"
-	conventions "go.opentelemetry.io/collector/model/semconv/v1.5.0"
+	conventions "go.opentelemetry.io/collector/model/semconv/v1.6.1"
 )
 
 // EncodeResourceMetadata encodes a metadata line from resource, writing to w.