+++ pkg/translator/jaeger/traces_to_jaegerproto.go
@@ -19,7 +19,7 @@ import (
 
 	"github.com/jaegertracing/jaeger/model"
 	"go.opentelemetry.io/collector/model/pdata"
-	conventions "go.opentelemetry.io/collector/model/semconv/v1.5.0"
+	conventions "go.opentelemetry.io/collector/model/semconv/v1.6.1"
 
 	"github.com/open-telemetry/opentelemetry-collector-contrib/internal/coreinternal/idutils"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/internal/coreinternal/tracetranslator"
@@ -445,7 +445,7 @@ func getTagsFromInstrumentationLibrary(il pdata.InstrumentationLibrary) ([]model
 	keyValues := make([]model.KeyValue, 0)
 	if ilName := il.Name(); ilName != "" {
 		kv := model.KeyValue{
-			Key:   conventions.InstrumentationLibraryName,
+			Key:   conventions.OtelLibraryName,
 			VStr:  ilName,
 			VType: model.ValueType_STRING,
 		}
@@ -453,7 +453,7 @@ func getTagsFromInstrumentationLibrary(il pdata.InstrumentationLibrary) ([]model
 	}
 	if ilVersion := il.Version(); ilVersion != "" {
 		kv := model.KeyValue{
-			Key:   conventions.InstrumentationLibraryVersion,
+			Key:   conventions.OtelLibraryVersion,
 			VStr:  ilVersion,
 			VType: model.ValueType_STRING,
 		}