+++ receiver/kubeletstatsreceiver/internal/kubelet/resource.go
@@ -18,7 +18,7 @@ import (
 	"fmt"
 
 	"go.opentelemetry.io/collector/model/pdata"
-	conventions "go.opentelemetry.io/collector/model/semconv/v1.5.0"
+	conventions "go.opentelemetry.io/collector/model/semconv/v1.6.1"
 	stats "k8s.io/kubelet/pkg/apis/stats/v1alpha1"
 )